+++ mayan/apps/tags/widgets.py
@@ -2,6 +2,15 @@ from django import forms
 from django.utils.html import conditional_escape
 
 
+class ColorWidget(forms.TextInput):
+    template_name = 'tags/widget_tag_color.html'
+
+    def __init__(self, attrs=None):
+        attrs = attrs or {}
+        attrs['type'] = 'color'
+        super().__init__(attrs=attrs)
+
+
 class TagFormWidget(forms.SelectMultiple):
     def __init__(self, *args, **kwargs):
         return super().__init__(*args, **kwargs)